session try

+++ src/Catty/EVCircularProgressView.h
@@ -49,7 +49,7 @@
 // A value from 0 to 1 that indicates how much progress has been made
 // When progress is zero, the progress view functions as an indeterminate progress indicator (a spinner)
 
-@property (nonatomic) float progress;
+@property (nonatomic) double progress;
 
 // On iOS 7, progressTintColor sets and gets the tintColor property, and therefore defaults to the value of tintColor
 // On iOS 6, defaults to [UIColor blackColor]
@@ -58,6 +58,6 @@
 
 // Set the value of the progress property, optionally animating the change
 
-- (void)setProgress:(float)progress animated:(BOOL)animated;
+- (void)setProgress:(double)progress animated:(BOOL)animated;
 
 @end
